@@ -238,7 +238,7 @@ func TestMultiValidator(t *testing.T) {
 		assert.NoError(t, err)
 
 		var wg sync.WaitGroup
-		for i := 0; i < 10; i++ {
+		for range 10 {
 			wg.Add(1)
 			go func() {
 				defer wg.Done()
