@@ -2,6 +2,7 @@ package workspace
 
 import (
 	"fmt"
+	"maps"
 	"sort"
 	"sync"
 
@@ -9,7 +10,6 @@ import (
 	"github.com/reearth/reearthx/i18n"
 	"github.com/reearth/reearthx/rerror"
 	"github.com/samber/lo"
-	"golang.org/x/exp/maps"
 )
 
 var (
